refactor(RetrofitRepository): RetrofitRepository룰 추가하여 뷰모델에서 데이터 조작 로직 분리
